LPS-65639 reorganizing code since at verify, the getAllPortlets() is not returning the needed data
