+++ modules/apps/foundation/server/server-admin-web/src/main/resources/META-INF/resources/resources.jsp
@@ -297,6 +297,15 @@ long usedMemory = totalMemory - runtime.freeMemory();
 					<aui:button cssClass="save-server-button" data-cmd="cleanUpPermissions" value="execute" />
 				</div>
 			</li>
+			<li class="clearfix list-group-item">
+				<div class="pull-left">
+					<h5><liferay-ui:message key="clean-up-portlet-preferences" /> <liferay-ui:icon-help message="clean-up-portlet-preferences-help" /></h5>
+				</div>
+
+				<div class="pull-right">
+					<aui:button cssClass="save-server-button" data-cmd="cleanUpPortletPreferences" value="execute" />
+				</div>
+			</li>
 		</ul>
 	</liferay-ui:panel>
 </liferay-ui:panel-container>
