+++ modules/apps/foundation/server/server-admin-web/src/main/resources/content/Language.properties
@@ -4,6 +4,8 @@ cancel-shutdown=Cancel Shutdown
 clean-up-actions=Clean Up Actions
 clean-up-permissions=Clean up permissions.
 clean-up-permissions-help=This process removes the assignment of some permissions on the Guest, User, and Power User roles in order to simplify the management of <em>User Customizable Pages</em>. Notably, <em>Add To Page</em> permissions is removed from the Guest, and User role for all portlets. Likewise, the same permission is reduced in scope for Power Users from portal wide to scoped to <em>User Personal Site</em>.
+clean-up-portlet-preferences=Clean Up Portlet Preferences
+clean-up-portlet-preferences-help=This process removes all orphaned Portlet Preferences that belong to Page Revisions. Any Portlet Preferences entry who's Portlet is not known to the Page Revision is removed. Portlet Preferences belonging to Runtime Portlets will also be removed.
 clear-content-cached-across-the-cluster=Clear content cached across the cluster.
 clear-content-cached-by-this-vm=Clear content cached by this VM.
 clear-the-database-cache=Clear the database cache.
