+++ portal-impl/src/com/liferay/portal/verify/VerifyProcessSuite.java
@@ -41,7 +41,6 @@ public class VerifyProcessSuite extends VerifyProcess {
 		verify(new VerifyLayout());
 		verify(new VerifyMessageBoards());
 		verify(new VerifyOrganization());
-		verify(new VerifyPortletPreferences());
 		verify(new VerifyRatings());
 		verify(new VerifyResourcePermissions());
 		verify(new VerifySocial());
